Ajustes a proyectos se agrego la validacion para mostrar los campos que son de la DGAA y se ajusto la parte de recursos financieros para que despliegue los decimales.
Tambien se hizo la correcion en Coautores Externos para que marque
los campos requeridos.

+++ app/DI.Colef.Sia.Core/RecursoFinancieroProyecto.cs
@@ -10,7 +10,7 @@ namespace DecisionesInteligentes.Colef.Sia.Core
         public virtual Institucion Institucion { get; set; }
 
         //[NotNullNotEmpty]
-        public virtual long Monto { get; set; }
+        public virtual decimal Monto { get; set; }
 
         [NotNull]
         public virtual Moneda Moneda { get; set; }
