+++ app/DI.Colef.Sia.Web.Controllers/Models/RecursoFinancieroProyectoForm.cs
@@ -8,7 +8,7 @@ namespace DecisionesInteligentes.Colef.Sia.Web.Controllers.Models
         public string Modificacion { get; set; }
         public bool Activo { get; set; }
 
-        public long Monto { get; set; }
+        public decimal Monto { get; set; }
 
         public int InstitucionId { get; set; }
         public string InstitucionNombre { get; set; }
